backend/requirements.txt
Django==5.2.2
djangorestframework==3.16.0
django-cors-headers==4.6.0
drf-spectacular==0.28.0
python-dotenv==1.1.0
gunicorn==23.0.0
PyJWT==2.10.1
cryptography==44.0.0
requests==2.32.3
openpyxl==3.1.5
pandas==2.2.3